+++ NEWS.md
@@ -1,5 +1,13 @@
 ## regexplain 0.1.x
 
+### 0.1.6
+
+* Added "Templates" button that opens a dialog to import regex patterns
+  (initially) sourced from <https://projects.lukehaas.me/regexhub>.
+* Added "Introduction", "Try These Examples" and "Resources" tabs to
+  the standard gadget Help tab.
+* Several bugfixes and tests were added around unicode character escaping within the gadget.
+
 ### 0.1.5
 
 * Options from **Regex** tab are extended to **Output** tab as well. Colored